longhui

+++ src/main/java/com/qg/servlet/llh/NoteCommentDelete.java
@@ -30,17 +30,27 @@ public class NoteCommentDelete extends HttpServlet {
 
 	protected void doGet(HttpServletRequest request, HttpServletResponse resp) throws ServletException, IOException {
 
-		int state = 501;
-		// 获取说说评论id
-		int commentId = Integer.parseInt(request.getParameter("commentId"));
-		//获取当前用户Id
-		int userId = ((UserModel) request.getSession().getAttribute("user")).getUserId();
-//		int userId =3;
-		// 删除服务器上的说说评论信息
-		if (!new NoteCommentService().deleteComment(commentId,userId)) {
+		int state = 501;//状态码
+		int userId = 0;//当前登陆用户
+		int commentId = 0;//评论id
+
+		/* 判断评论对象是否为null */
+		if (request.getParameter("commentId").equals(null)||request.getParameter("commentId").equals("")) {
 			state = 502;
+			LOGGER.log(Level.ERROR, " 留言评论删除出现空指针");
+		} else {
+			// 获取说说评论id
+			commentId = Integer.parseInt(request.getParameter("commentId"));
+			// 获取当前用户Id
+			userId = ((UserModel) request.getSession().getAttribute("user")).getUserId();
+
+			// 删除服务器上的说说评论信息
+			if (!new NoteCommentService().deleteComment(commentId, userId)) {
+				state = 502;
+			}
 		}
-		LOGGER.log(Level.DEBUG, " {0}想删除留言评论，其id为{1},状态为", userId,commentId,state);
+		LOGGER.log(Level.DEBUG, " {0}想删除留言评论，其id为{1},状态为", userId, commentId, state);
+
 		DataOutputStream output = new DataOutputStream(resp.getOutputStream());
 		output.write(JsonUtil.tojson(state).getBytes("UTF-8"));
 		output.close();
